+++ CHANGES.rst
@@ -56,6 +56,8 @@ Improvements
 - When blocking a user account, the user will be forcefully logged out in
   addition to being prevented from logging in
 - Warn when editing a speaker/author would result in duplicate emails
+- Show track-related columns in abstract list only if there are tracks
+  defined for the event (:issue:`2813`)
 
 Bugfixes
 ^^^^^^^^